add fusion mutator
