+++ src/Nncase.Core/IR/NN/Conv2D.cs
@@ -57,5 +57,11 @@ namespace Nncase.IR.NN
         /// Gets FusedClamp.
         /// </summary>
         public static readonly ParameterInfo FusedClamp = new(typeof(Conv2D), 7, "fused_clamp", HasShape(new Shape(2)) & HasDataType(DataTypes.Float32));
+
+        /// <inheritdoc/>
+        public override string DisplayProperty()
+        {
+            return $"PadMode: {PadMode}";
+        }
     }
 }
