+++ src/Nncase.EGraph/Transform/EGraphRewriter.cs
@@ -56,7 +56,7 @@ internal static class EGraphRewriter
                     }
 
                     var newEClass = eGraph.Add(newExpr);
-                    if (options.DumpLevel > 3)
+                    if (options.DumpLevel > 4)
                     {
                         Console.WriteLine($"Version {eGraph.Version} : Merge {{{oldEClass}}} to {{{newEClass}}}");
                     }
